backend/requirements.txt
fastapi
uvicorn
python-multipart
websockets
numpy
scipy
firebase-admin
openai
python-dotenv
